Extract player routes to own file
